requirements.txt
PTable==0.9.2
requests==2.22.0
tabulate==0.8.7
validate_email==1.3
google==2.0.3
validators==0.14.3
xlwt==1.3.0
auto_mix_prep==0.2.0
beautifulsoup4==4.9.0
dnspython==1.16.0
mysql_connector_repackaged==0.3.1
prettytable==0.7.2
mysql-connector-python==8.0.19
